requirements.txt
aiofiles==23.2.1
Django==4.2.7
httpx==0.25.2
psycopg2==2.8.6
celery[redis]==5.2.7
flower==1.2.0
django-celery-beat==2.5.0
django-environ==0.11.2

